requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
python-dotenv>=1.0.0
lxml>=4.9.0

